refactor:  정리 작업 - app.py에서 불필요한 코드 정리 및 모듈 임포트 순서 변경. - main.py에서 주문 라우터 경로를 수정하여 일관성 유지.
