+++ main.py
@@ -14,7 +14,7 @@ from contextlib import asynccontextmanager
 
 from fastapi import FastAPI, APIRouter
 from fastapi.middleware.cors import CORSMiddleware
-from api.v1.endpoints.order import router as order_router
+from api.v1.endpoints.receive_order import router as order_router
 from api.v1.endpoints.product import router as products_router
 from api.v1.endpoints.mall_price import router as mall_price_router
 from utils.logs.sabangnet_logger import get_logger, HTTPLoggingMiddleware
